chore: add fake-channel-client
