+++ packages/channel-client/package.json
@@ -22,6 +22,9 @@
   "author": "snario <liam@l4v.io>",
   "license": "MIT",
   "devDependencies": {
+    "@types/eslint": "^6.1.3",
+    "@types/eslint-plugin-prettier": "^2.2.0",
+    "@types/prettier": "^1.19.0",
     "@typescript-eslint/eslint-plugin": "^2.10.0",
     "@typescript-eslint/parser": "^2.10.0",
     "eslint": "^6.6.0",
