+++ packages/channel-client/src/utils.ts
@@ -1,0 +1,3 @@
+export function sleep(time: number) {
+  return new Promise(resolve => setTimeout(resolve, time));
+}
